Auditlogituksen korjaus myös Hakemuspalvelun servletiin
