+++ src/main/scala/fi/oph/koski/hakemuspalvelu/HakemuspalveluService.scala
@@ -16,29 +16,8 @@ class HakemuspalveluService(application: KoskiApplication) extends GlobalExecuti
 
   def findOppija(oppijaOid: String)
     (implicit koskiSession: KoskiSpecificSession): Either[HttpStatus, HakemuspalveluOppija] = {
-
-    val oppija = opiskeluoikeusFacade.haeOpiskeluoikeudet(oppijaOid, HakemuspalveluSchema.schemassaTuetutOpiskeluoikeustyypit, useDownloadedYtr = true)
+    opiskeluoikeusFacade.haeOpiskeluoikeudet(oppijaOid, HakemuspalveluSchema.schemassaTuetutOpiskeluoikeustyypit, useDownloadedYtr = true)
       .map(teePalautettavaHakemuspalveluOppija)
-
-    oppija.map(o => o.opiskeluoikeudet.foreach {
-      case x: HakemuspalveluDIAOpiskeluoikeus if x.oid.isDefined => auditLog(o.henkilö.oid, opiskeluoikeusOid = x.oid.get)
-      case x: HakemuspalveluEBTutkinnonOpiskeluoikeus if x.oid.isDefined => auditLog(o.henkilö.oid, opiskeluoikeusOid = x.oid.get)
-      case x: HakemuspalveluYlioppilastutkinnonOpiskeluoikeus if x.oid.isDefined => auditLog(o.henkilö.oid, opiskeluoikeusOid = x.oid.get)
-      case _ => // Do nothing for other types
-    })
-
-    oppija
-  }
-
-  def findOppijaByHetu(hetu: String)
-    (implicit koskiSession: KoskiSpecificSession): Either[HttpStatus, HakemuspalveluOppija] = {
-
-    val oppijaResult = application.opintopolkuHenkilöFacade.findOppijaByHetu(hetu)
-
-    oppijaResult match {
-      case Some(o) => findOppija(o.oid)
-      case None => Left(KoskiErrorCategory.notFound.oppijaaEiLöydyHetulla())
-    }
   }
 
   private def teePalautettavaHakemuspalveluOppija(
@@ -98,17 +77,4 @@ class HakemuspalveluService(application: KoskiApplication) extends GlobalExecuti
       => true
     }
   }
-
-  private def auditLog(oppijaOid: String, opiskeluoikeusOid: String)(implicit user: KoskiSpecificSession): Unit =
-    AuditLog
-      .log(
-        KoskiAuditLogMessage(
-          KoskiOperation.HAKEMUSPALVELU_OPISKELUOIKEUS_HAKU,
-          user,
-          Map(
-            KoskiAuditLogMessageField.oppijaHenkiloOid -> oppijaOid,
-            KoskiAuditLogMessageField.opiskeluoikeusOid -> opiskeluoikeusOid,
-          )
-        )
-      )
 }
